+++ examples/sqlmodel/tests/data/__init__.py
@@ -2,3 +2,4 @@ from .user import USER_DATA
 from .role import ROLE_DATA
 from .company import COMPANY_DATA
 from .project import PROJECT_DATA
+from .zone import ZONE_DATA