+++ examples/sqlmodel/tests/data/zone.py
@@ -1,0 +1,20 @@
+ZONE_DATA = [
+    {
+        "id": 1,
+        "name": "zone1",
+        "description": "zone1 description",
+        "is_active": True
+    },
+    {
+        "id": 2,
+        "name": "zone2",
+        "description": "Project2 description",
+        "is_active": True
+    },
+    {
+        "id": 3,
+        "name": "zone3",
+        "description": "zone3 description",
+        "is_active": True
+    }
+]